requirements/base.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-dotenv>=1.0.0

# Database clients
google-cloud-bigquery>=3.11.4
redis>=5.0.1

# CORS
starlette>=0.27.0 